requirements.txt
aiohttp==3.8.5
aioresponses==0.7.4
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
httpretty==1.1.4
idna==3.4
multidict==6.0.4
python-Wappalyzer==0.3.1
requests==2.31.0
soupsieve==2.4.1
termcolor==2.3.0
urllib3==2.0.4
yarl==1.9.2
